libpanic_unwind: Unconfigure tests during normal build
