+++ src/libpanic_unwind/dwarf/mod.rs
@@ -6,6 +6,9 @@
 // are compiling it everywhere to avoid regressions.
 #![allow(unused)]
 
+#[cfg(test)]
+mod tests;
+
 pub mod eh;
 
 use core::mem;
@@ -66,23 +69,5 @@ impl DwarfReader {
             result |= (!0 as u64) << shift;
         }
         result as i64
-    }
-}
-
-#[test]
-fn dwarf_reader() {
-    let encoded: &[u8] = &[1, 2, 3, 4, 5, 6, 7, 0xE5, 0x8E, 0x26, 0x9B, 0xF1, 0x59, 0xFF, 0xFF];
-
-    let mut reader = DwarfReader::new(encoded.as_ptr());
-
-    unsafe {
-        assert!(reader.read::<u8>() == u8::to_be(1u8));
-        assert!(reader.read::<u16>() == u16::to_be(0x0203));
-        assert!(reader.read::<u32>() == u32::to_be(0x04050607));
-
-        assert!(reader.read_uleb128() == 624485);
-        assert!(reader.read_sleb128() == -624485);
-
-        assert!(reader.read::<i8>() == i8::to_be(-1));
     }
 }
